% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_linear_rating_curve}
\alias{calculate_linear_rating_curve}
\title{Calculate Linear Rating Curve}
\usage{
calculate_linear_rating_curve(site_number, start_date, end_date)
}
\arguments{
\item{site_number}{USGS stream gage site number.}

\item{start_date}{Start date, provide as \code{"YYYY-mm-dd"}.}

\item{end_date}{End date, provide as \code{"YYYY-mm-dd"}.}
}
\value{
Data frame with slope and intercept for each site number provided.
Values are in imperial units.
}
\description{
Calculate linear rating curve for USGS stream gage site between
a given start and end date
}
