% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{fetch_gage_data}
\alias{fetch_gage_data}
\title{Fetch Flow Data from USGS Stream Gage}
\usage{
fetch_gage_data(site_number, start_date, end_date)
}
\arguments{
\item{site_number}{USGS stream gage site number.}

\item{start_date}{Start date, provide as \code{"YYYY-mm-dd"}.}

\item{end_date}{End date, provide as \code{"YYYY-mm-dd"}.}
}
\value{
A list of data frames. \code{site_data} is the time series flow data
in local date time and cubic feet per second. \code{site_metadata} is the metadata
associated with the requested gage station.
}
\description{
Fetch flow data from USGS stream gage
}
