% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factor_data.R
\name{get_factor_data}
\alias{get_factor_data}
\title{Get factor data}
\usage{
get_factor_data()
}
\value{

}
\description{
This function collects data from Fama and French's five factor database. It retrieves daily factors for developed markets.
}
