% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_data.R
\name{get_stock_data}
\alias{get_stock_data}
\title{Get stock data}
\usage{
get_stock_data(tickers)
}
\arguments{
\item{tickers}{}
}
\value{

}
\description{
This function collects market data for the specified tickers and computes both simple and logarithmic returns.
}
\examples{

}
