\name{matchedPortfolioCollection-class}
\docType{class}
\alias{matchedPortfolioCollection-class}
\alias{summary,matchedPortfolioCollection-method}
\alias{plot,matchedPortfolioCollection,missing-method}
\alias{matching,data.frame-method}

\title{Class "matchedPortfolioCollection"}
\description{ A collection of objects of class \code{matchedPortfolio}.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("matchedPortfolioCollection", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{data}:}{Object of class \code{"list"} A list of objects
      of class \code{matchedPortfolio}. }
  }
}
\section{Methods}{
No methods defined with class "matchedPortfolioCollection" in the signature.
}

\author{Jeff Enos \email{jeff@kanecap.com}}
\keyword{classes}
