% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factor_returns_cum.R
\name{plot_factor_returns_cum}
\alias{plot_factor_returns_cum}
\title{Plot factor returns cumulative}
\usage{
plot_factor_returns_cum(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
This function plots daily cumulative returns of market factors.
}
