% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rolling_stat.R
\name{plot_rolling_stat}
\alias{plot_rolling_stat}
\title{Plot rolling stat}
\usage{
plot_rolling_stat(.data, ticker, width = 200)
}
\arguments{
\item{.data}{}

\item{ticker}{}

\item{width}{length (days) of rolling stat}
}
\value{

}
\description{
This function plot returns with rolling mean and median
}
\details{
TODO: add labels for easier use with purrr::map
}
