% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaled_returns.R
\name{plot_scaled_returns}
\alias{plot_scaled_returns}
\title{Plot scaled returns}
\usage{
plot_scaled_returns(
  data,
  id = symbol,
  price_var = adjusted,
  start_date,
  date_var = date
)
}
\arguments{
\item{data}{}

\item{id}{}

\item{price_var}{}

\item{start_date}{}

\item{date_var}{}
}
\value{

}
\description{
Todo: create tests
}
