% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_volatility.R
\name{plot_stock_volatility}
\alias{plot_stock_volatility}
\title{Plot stock volatility}
\usage{
plot_stock_volatility(.data, ticker, return_var, all_tickers = FALSE)
}
\arguments{
\item{.data}{}

\item{ticker}{}

\item{all_tickers}{boolean}
}
\value{

}
\description{
This function plots the daily stock returns
}
