\name{portfolio-package}
\alias{portfolio-package}
\alias{portfolio}
\docType{package}
\title{
Analysing equity portfolios
}
\description{
Classes for analysing and implementing equity portfolios.
}
\details{
\tabular{ll}{
Package: \tab portfolio\cr
Version: \tab 0.3\cr
Date: \tab 2006-04-28\cr
Depends: \tab R (>= 2.0.0), methods, graphics, grid, lattice\cr
License: \tab GPL version 2 or later.\cr
Built: \tab R 2.3.0; ; 2006-06-01 17:57:45; unix\cr
}

Index:
\preformatted{
assay                   Assay Research rankings as of 2004-12-31
contribution-class      Class "contribution"
contributionHistory-class
                        Class "contributionHistory"
dow.jan.2005            DJIA for January, 2005
exposure-class          Class "exposure"
exposureHistory-class   Class "exposureHistory"
global.2004             Security data of large global companies for
                        2004
global.2004.history     portfolioHistory object formed from the
                        global.2004 dataset
objectHistory-class     Class "objectHistory"
performance-class       Class "performance"
performanceHistory-class
                        Class "performanceHistory"
portfolio-class         Class "portfolio"
portfolioBasic-class    Class "portfolioBasic"
portfolioHistory-class
                        Class "portfolioHistory"
tradelist-class         Class "tradelist"
weight                  Calculate Position Weights
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{matching_portfolio} \tab Matching Portfolios (source, pdf)\cr
\code{portfolio} \tab Using the portfolio package (source, pdf)\cr
\code{tradelist} \tab Using the tradelist class (source, pdf)\cr
}
An overview will be provided soon.
}
\author{
Jeff Enos <jeff@kanecap.com> and David Kane <dave@kanecap.com>,
with contributions from Daniel Gerlanc <daniel@gerlanc.com>

Maintainer: Jeff Enos <jeff@kanecap.com>
}
\keyword{ package }
