% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-portfolio.R
\name{render_portfolio}
\alias{render_portfolio}
\title{Render Portfolio}
\usage{
render_portfolio(
  site_number = NULL,
  state = NULL,
  start_date = NULL,
  end_date = NULL,
  shiny = FALSE
)
}
\arguments{
\item{site_number}{Optional. A USGS stream gage number. Explore stream gages on NWIS
\link{https://maps.waterdata.usgs.gov/mapper/index.html}. If \code{NULL}, a random
eligible site is selected. Do not provide if state is specified.}

\item{state}{Optional. A two letter abbreviation for a US state, e.g. \code{"CO"}.
Only provide if \code{site_number} is not given.}

\item{start_date}{Optional. Provide as \code{"YYYY-mm_dd"}, otherwise 1 month
prior to today will be used.}

\item{end_date}{Optional. Provide as \code{"YYYY-mm-dd"}, otherwise today will be used.}

\item{shiny}{Logical. Should the portfolio be rendered as an R Shiny app?
Default is \code{FALSE} which renders the portfolio in Rmarkdown format.}
}
\value{
Portfolio in Rmarkdown or Shiny format.
}
\description{
Render portfolio.
}
