% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{fetch_weather_data}
\alias{fetch_weather_data}
\title{Fetch Weather Data From Open Meteo}
\usage{
fetch_weather_data(site_metadata, start_date, end_date)
}
\arguments{
\item{site_metadata}{A data frame containing the columns \code{site_number,
lat, lon} where \code{site_number} corresponds to the USGS site number,
\code{lat} to the latitudinal coordinate of the given site and \code{lon}
to the longitudinal coordinate of the given site.}

\item{start_date}{Start date, provide as \code{"YYYY-mm-dd"}.}

\item{end_date}{End date, provide as \code{"YYYY-mm-dd"}.}
}
\value{
A data frame containing a time series of precipitation data
in local date time and inches.
}
\description{
Fetch precipitation data from Open Meteo
\link{https://open-meteo.com/en/docs/historical-weather-api#api_form}
}
