% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_returns_distribution.R
\name{plot_returns_distribution}
\alias{plot_returns_distribution}
\title{Plot returns distribution}
\usage{
plot_returns_distribution(.data, ticker, return_var, all_tickers = FALSE)
}
\arguments{
\item{.data}{}

\item{return_var}{}

\item{all_tickers}{boolean}
}
\value{

}
\description{
This function plots the distribution of all symbols in data.
}
\examples{

}
