\name{portfolio-package}
\alias{portfolio-package}
\alias{portfolio}
\docType{package}
\title{
Analysing equity portfolios
}
\description{
Classes for analysing and implementing equity portfolios.
}
\details{
\tabular{ll}{
Package: \tab portfolio\cr
Version: \tab 0.3-1\cr
Date: \tab 2006-10-04\cr
Depends: \tab R (>= 2.4.0), methods, graphics, grid, lattice, nlme\cr
License: \tab GPL version 2 or later.\cr
Packaged: \tab Wed Oct 4 09:35:05 2006; enos\cr
Built: \tab R 2.4.0; ; 2006-10-04 10:00:34; unix\cr
}

Index:
\preformatted{
assay                   Assay Research rankings as of 2004-12-31
contribution-class      Class "contribution"
dow.jan.2005            DJIA for January, 2005
exposure-class          Class "exposure"
global.2004             Security data of large global companies for
                        2004
map.market              Create a Map of the Market
matchedPortfolio-class
                        Class "matchedPortfolio"
performance-class       Class "performance"
portfolio-class         Class "portfolio"
portfolio-package       Analysing equity portfolios
portfolioBasic-class    Class "portfolioBasic"
tradelist-class         Class "tradelist"
trades-class            Class "trades"
weight                  Calculate Position Weights
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{matching_portfolio} \tab Matching Portfolios (source, pdf)\cr
\code{portfolio} \tab Using the portfolio package (source, pdf)\cr
\code{tradelist} \tab Using the tradelist class (source, pdf)\cr
}
}
\author{
Jeff Enos <jeff@kanecap.com> and David Kane <dave@kanecap.com>,
with contributions from Daniel Gerlanc <daniel@gerlanc.com> and
Kyle Campbell <Kyle.W.Campbell@williams.edu>

Maintainer: Jeff Enos <jeff@kanecap.com>
}
\keyword{ package }
